Try all constructors/methods rather than searching
Instead of trying to infer an array of parameter types to allow
constructors and methods to be found using Class.getConstructor and
Class.getMethod, now just get all of them using Class.getConstructors
and Class.getMethods and attempt to invoke each in turn, where, in
the case of methods, the method name matches.  The first attempt
which doesn't throw an IllegalArgumentException is assumed to be the
desired constructor or method.

This new procedure should be more robust and also allows for
polymorphism since we no longer require exact parameter type
matches.